[0.18.0-SNAPSHOT] NativeImageOptions added All static util methods moved out of AbstractHintProcessor to HintUtils
